+++ bindings/nodejs/examples/21-claim-outputs.js
@@ -23,9 +23,7 @@ async function run() {
                 recipientAddress,
                 amount,
                 unlocks: {
-                    expiration: {
-                        unixTime: Math.round(new Date().getTime() / 1000) + 15000
-                    }
+                    expirationUnixTime: Math.round(new Date().getTime() / 1000) + 15000
                 }
             },
         );
@@ -43,14 +41,18 @@ async function run() {
 }
 
 async function handleNewOutputOfBob(err, data) {
-    console.log('Output received:', data)
-    const event = JSON.parse(data)
-    if (event.accountIndex === bob.meta.index) {
-        const outputId = event.event.NewOutput.output.outputId
-        await bob.sync()
-        const resp = await bob.claimOutputs([outputId])
-        console.log('Output has been claimed in the following transaction:', resp)
-        process.exit(0)
+    try {
+        console.log('Output received:', data)
+        const event = JSON.parse(data)
+        if (event.accountIndex === bob.meta.index) {
+            const outputId = event.event.NewOutput.output.outputId
+            await bob.sync()
+            const resp = await bob.claimOutputs([outputId])
+            console.log('Output has been claimed in the following transaction:', resp)
+            process.exit(0)
+        }
+    } catch (error) {
+        console.log('Error: ', error);
     }
 }
 run();