+++ bindings/nodejs/types/account.ts
@@ -15,7 +15,7 @@ export type AccountId = number | string;
 
 /** The balance of an account */
 export interface AccountBalance {
-    /**  The balance of the base coin */
+    /** The balance of the base coin */
     baseCoin: BaseCoinBalance;
     /** The required storage deposit for the outputs */
     requiredStorageDeposit: RequiredStorageDeposit;
@@ -84,7 +84,7 @@ export interface SyncOptions {
     /// Try to sync transactions from incoming outputs with their inputs. Some data may not be obtained if it has been
     /// pruned.
     syncIncomingTransactions?: boolean;
-    /** Checks pending transactions and promotes/reattaches them if necessary.  Default: true. */
+    /** Checks pending transactions and promotes/reattaches them if necessary. Default: true. */
     syncPendingTransactions?: boolean;
     /** Specifies what outputs should be synced for the ed25519 addresses from the account. */
     account?: AccountSyncOptions;
@@ -92,7 +92,7 @@ export interface SyncOptions {
     alias?: AliasSyncOptions;
     /** Specifies what outputs should be synced for the address of an nft output. */
     nft?: NftSyncOptions;
-    /** Specifies if only basic outputs with an AddressUnlockCondition alone should be synced, will overwrite `account`, `alias` and `nft` options.  Default: false. */
+    /** Specifies if only basic outputs with an AddressUnlockCondition alone should be synced, will overwrite `account`, `alias` and `nft` options. Default: false. */
     syncOnlyMostBasicOutputs?: boolean;
     /** Sync native token foundries, so their metadata can be returned in the balance. Default: false. */
     syncNativeTokenFoundries?: boolean;